pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bd_transformer"
version = "0.1.0"
description = "A take-home assignment for ML DE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet",
    "numpy",
    "pandas",
    "pyarrow",
    "pyyaml",
    "pyspark>=3.4.0",
    "apache-airflow>=2.7.0",
    "psutil",
    "matplotlib"
]

[tool.setuptools.packages.find]
include = ["bd_transformer*"]
exclude = ["data*"]
